requirements.txt
Flask==3.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.4
WTForms==3.1.2
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
PyMySQL==1.1.1
mysqlclient==2.2.4
email_validator==2.2.0
python-dotenv==1.0.1
pillow==11.2.1
beautifulsoup4==4.13.4
requests==2.32.3
python-dateutil==2.9.0.post0
